+++ host/Makefile
@@ -3,7 +3,7 @@ CFLAGS = -std=c++11 -g -Wall -Werror -O3
 
 LDLIBS = -lcrypto
 
-SRC = host_sha256.cpp
+SRC = host_sha256.cpp host_benchmark.cpp
 BIN = host_sha256
 
 OBJ = $(SRC:.c=.o)
