--- host/sha256.hpp
@@ -1,8 +1,0 @@
-#ifndef __SHA256_H_INCLUDED__
-#define __SHA256_H_INCLUDED__
-
-#include <string.h>
- int SHA256_CPU_debug(std::string filename);
- std::vector<std::string> SHA256_CPU_verify(std::string filename);
-
- #endif
